creating models and tests packages
