+++ models/youtube_service_model.go
@@ -1,0 +1,18 @@
+package models
+
+type PlaylistItemsResponse struct {
+	Items []PlaylistItem `json:"items"`
+}
+
+type PlaylistItem struct {
+	Snippet VideoSnippet `json:"snippet"`
+}
+
+type VideoSnippet struct {
+	ResourceId  ResourceId `json:"resourceId"`
+	Description string     `json:"description"`
+}
+
+type ResourceId struct {
+	VideoId string `json:"videoId"`
+}
