+++ tests/description_to_links_mapper_test.go
@@ -1,4 +1,4 @@
-package main
+package tests
 
 import "testing"
