+++ tests/youtube_service_test.go
@@ -1,4 +1,4 @@
-package main
+package tests
 
 import (
 	"testing"
